refactor: 게시글 목록, 상세에서 좋아요 판별 추가
- 좋아요 기능에서 이미 좋아요 한 게시글 검증 추가
- 좋아요 API 호출 시 응답값 수정 (vodi -> boolean)
- swagger 호출 시 access token 입력 추가

fix: 모임 등록 시 프로필 사진 등록 제거

- 테스트 코드 수정

+++ src/main/java/com/assemble/post/service/PostLikeService.java
@@ -19,24 +19,45 @@ public class PostLikeService {
     private final PostRepository postRepository;
 
     @Transactional(rollbackFor = AssembleException.class)
-    public void createPostLike(PostLikeRequest postLikeRequest) {
+    public boolean createPostLike(PostLikeRequest postLikeRequest) {
         Likes postLike = postLikeRequest.toEntity();
+        if (isAleadyLikeByUser(postLikeRequest)) {
+            return false;
+        }
+
         postLikeRepository.save(postLike);
 
         Post post = postRepository.findById(postLikeRequest.getPostId())
                 .orElseThrow(() -> new NotFoundException(Post.class, postLikeRequest.getPostId()));
 
         post.increaseLikes();
+
+        return true;
     }
 
     @Transactional(rollbackFor = AssembleException.class)
-    public void deletePostLike(PostLikeRequest postLikeRequest) {
+    public boolean deletePostLike(PostLikeRequest postLikeRequest) {
         Likes postLike = postLikeRequest.toEntity();
+        if (!isAleadyLikeByUser(postLikeRequest)) {
+            return false;
+        }
+
         postLikeRepository.delete(postLike);
 
         Post post = postRepository.findById(postLikeRequest.getPostId())
                 .orElseThrow(() -> new NotFoundException(Post.class, postLikeRequest.getPostId()));
 
         post.decreaseLikes();
+
+        return true;
+    }
+
+    @Transactional(readOnly = true)
+    public boolean isAleadyLikeByUser(PostLikeRequest postLikeRequest) {
+        if (postLikeRepository.findPostByUser(postLikeRequest).isPresent()) {
+            return true;
+        }
+
+        return false;
     }
 }
